+++ lib/android.dart
@@ -1,10 +1,12 @@
+// ignore_for_file: public_member_api_docs
+
 import 'dart:io';
 
+import 'package:flutter_launcher_icons/config/config.dart';
 import 'package:flutter_launcher_icons/constants.dart';
 import 'package:flutter_launcher_icons/constants.dart' as constants;
 import 'package:flutter_launcher_icons/custom_exceptions.dart';
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart';
-import 'package:flutter_launcher_icons/utils.dart';
+import 'package:flutter_launcher_icons/utils.dart' as utils;
 import 'package:flutter_launcher_icons/xml_templates.dart' as xml_template;
 import 'package:image/image.dart';
 import 'package:path/path.dart' as path;
@@ -33,23 +35,23 @@ List<AndroidIconTemplate> androidIcons = <AndroidIconTemplate>[
 ];
 
 void createDefaultIcons(
-  FlutterLauncherIconsConfig flutterLauncherIconsConfig,
+  Config config,
   String? flavor,
 ) {
-  printStatus('Creating default icons Android');
-  // todo: support prefixPath
-  final String? filePath = flutterLauncherIconsConfig.getImagePathAndroid();
+  utils.printStatus('Creating default icons Android');
+  // TODO(p-mazhnik): support prefixPath
+  final String? filePath = config.getImagePathAndroid();
   if (filePath == null) {
     throw const InvalidConfigException(errorMissingImagePath);
   }
-  final Image? image = decodeImageFile(filePath);
+  final Image? image = utils.decodeImageFile(filePath);
   if (image == null) {
     return;
   }
   final File androidManifestFile = File(constants.androidManifestFile);
-  if (flutterLauncherIconsConfig.isCustomAndroidFile) {
-    printStatus('Adding a new Android launcher icon');
-    final String iconName = flutterLauncherIconsConfig.android;
+  if (config.isCustomAndroidFile) {
+    utils.printStatus('Adding a new Android launcher icon');
+    final String iconName = config.android;
     isAndroidIconNameCorrectFormat(iconName);
     final String iconPath = '$iconName.png';
     for (AndroidIconTemplate template in androidIcons) {
@@ -57,7 +59,7 @@ void createDefaultIcons(
     }
     overwriteAndroidManifestWithNewLauncherIcon(iconName, androidManifestFile);
   } else {
-    printStatus(
+    utils.printStatus(
       'Overwriting the default Android launcher icon with a new icon',
     );
     for (AndroidIconTemplate template in androidIcons) {
@@ -87,20 +89,18 @@ bool isAndroidIconNameCorrectFormat(String iconName) {
 }
 
 void createAdaptiveIcons(
-  FlutterLauncherIconsConfig flutterLauncherIconsConfig,
+  Config config,
   String? flavor,
 ) {
-  printStatus('Creating adaptive icons Android');
+  utils.printStatus('Creating adaptive icons Android');
 
   // Retrieve the necessary Flutter Launcher Icons configuration from the pubspec.yaml file
-  final String? backgroundConfig =
-      flutterLauncherIconsConfig.adaptiveIconBackground;
-  final String? foregroundImagePath =
-      flutterLauncherIconsConfig.adaptiveIconForeground;
+  final String? backgroundConfig = config.adaptiveIconBackground;
+  final String? foregroundImagePath = config.adaptiveIconForeground;
   if (backgroundConfig == null || foregroundImagePath == null) {
     throw const InvalidConfigException(errorMissingImagePath);
   }
-  final Image? foregroundImage = decodeImageFile(foregroundImagePath);
+  final Image? foregroundImage = utils.decodeImageFile(foregroundImagePath);
   if (foregroundImage == null) {
     return;
   }
@@ -118,12 +118,12 @@ void createAdaptiveIcons(
   // Create adaptive icon background
   if (isAdaptiveIconConfigPngFile(backgroundConfig)) {
     _createAdaptiveBackgrounds(
-      flutterLauncherIconsConfig,
+      config,
       backgroundConfig,
       flavor,
     );
   } else {
-    createAdaptiveIconMipmapXmlFile(flutterLauncherIconsConfig, flavor);
+    createAdaptiveIconMipmapXmlFile(config, flavor);
     updateColorsXmlFile(backgroundConfig, flavor);
   }
 }
@@ -138,11 +138,13 @@ void createAdaptiveIcons(
 void updateColorsXmlFile(String backgroundConfig, String? flavor) {
   final File colorsXml = File(constants.androidColorsFile(flavor));
   if (colorsXml.existsSync()) {
-    printStatus('Updating colors.xml with color for adaptive icon background');
+    utils.printStatus(
+      'Updating colors.xml with color for adaptive icon background',
+    );
     updateColorsFile(colorsXml, backgroundConfig);
   } else {
-    printStatus('No colors.xml file found in your Android project');
-    printStatus(
+    utils.printStatus('No colors.xml file found in your Android project');
+    utils.printStatus(
       'Creating colors.xml file and adding it to your Android project',
     );
     createNewColorsFile(backgroundConfig, flavor);
@@ -152,14 +154,12 @@ void updateColorsXmlFile(String backgroundConfig, String? flavor) {
 /// Creates the xml file required for the adaptive launcher icon
 /// FILE LOCATED HERE: res/mipmap-anydpi/{icon-name-from-yaml-config}.xml
 void createAdaptiveIconMipmapXmlFile(
-  FlutterLauncherIconsConfig flutterLauncherIconsConfig,
+  Config config,
   String? flavor,
 ) {
-  if (flutterLauncherIconsConfig.isCustomAndroidFile) {
+  if (config.isCustomAndroidFile) {
     File(
-      constants.androidAdaptiveXmlFolder(flavor) +
-          flutterLauncherIconsConfig.android +
-          '.xml',
+      constants.androidAdaptiveXmlFolder(flavor) + config.android + '.xml',
     ).create(recursive: true).then((File adaptiveIcon) {
       adaptiveIcon.writeAsString(xml_template.icLauncherXml);
     });
@@ -176,12 +176,12 @@ void createAdaptiveIconMipmapXmlFile(
 
 /// creates adaptive background using png image
 void _createAdaptiveBackgrounds(
-  FlutterLauncherIconsConfig flutterLauncherIconsConfig,
+  Config config,
   String adaptiveIconBackgroundImagePath,
   String? flavor,
 ) {
   final String filePath = adaptiveIconBackgroundImagePath;
-  final Image? image = decodeImageFile(filePath);
+  final Image? image = utils.decodeImageFile(filePath);
   if (image == null) {
     return;
   }
@@ -199,11 +199,9 @@ void _createAdaptiveBackgrounds(
 
   // Creates the xml file required for the adaptive launcher icon
   // FILE LOCATED HERE:  res/mipmap-anydpi/{icon-name-from-yaml-config}.xml
-  if (flutterLauncherIconsConfig.isCustomAndroidFile) {
+  if (config.isCustomAndroidFile) {
     File(
-      constants.androidAdaptiveXmlFolder(flavor) +
-          flutterLauncherIconsConfig.android +
-          '.xml',
+      constants.androidAdaptiveXmlFolder(flavor) + config.android + '.xml',
     ).create(recursive: true).then((File adaptiveIcon) {
       adaptiveIcon.writeAsString(xml_template.icLauncherDrawableBackgroundXml);
     });
@@ -266,7 +264,7 @@ void overwriteExistingIcons(
   String filename,
   String? flavor,
 ) {
-  final Image newFile = createResizedImage(template.size, image);
+  final Image newFile = utils.createResizedImage(template.size, image);
   File(
     constants.androidResFolder(flavor) +
         template.directoryName +
@@ -286,7 +284,7 @@ void _saveNewImages(
   String iconFilePath,
   String? flavor,
 ) {
-  final Image newFile = createResizedImage(template.size, image);
+  final Image newFile = utils.createResizedImage(template.size, image);
   File(
     constants.androidResFolder(flavor) +
         template.directoryName +